Create Discovery and Reconstitution of the Cycloclavine Biosynthetic Pathway—Enzymatic Formation of a Cyclopropyl Group
The ergot alkaloids, a class of fungal-derived natural products with important biological activities, are derived from a common intermediate, chanoclavine-I, which is elaborated into a set of diverse structures. Herein we report the discovery of the biosynthetic pathway of cycloclavine, a complex ergot alkaloid containing a cyclopropyl moiety. We used a yeast-based expression platform along with in vitro biochemical experiments to identify the enzyme that catalyzes a rearrangement of the chanoclavine-I intermediate to form a cyclopropyl moiety. The resulting compound, cycloclavine, was produced in yeast at titers of >500 mg L−1, thus demonstrating the feasibility of the heterologous expression of these complex alkaloids.