Create The Dsl1 Tethering Complex Actively Participates in Soluble NSF (N-Ethylmaleimide-sensitive Factor) Attachment Protein Receptor (SNARE) Complex Assembly at the Endoplasmic Reticulum in Saccharomyces cerevisiae
Intracellular transport is largely dependent on vesicles that bud off from one compartment and fuse with the target compartment. The first contact of an incoming vesicle with the target membrane is mediated by tethering factors. The tethering factor responsible for recruiting Golgi-derived vesicles to the ER is the Dsl1 tethering complex, which is comprised of the essential proteins Dsl1p, Dsl3p, and Tip20p. We investigated the role of the Tip20p subunit at the ER by analyzing two mutants, tip20-5 and tip20-8. Both mutants contained multiple mutations that were scattered throughout the TIP20 sequence. Individual mutations could not reproduce the temperature-sensitive phenotype of tip20-5 and tip20-8, indicating that the overall structure of Tip20p might be altered in the mutants. Using molecular dynamics simulations comparing Tip20p and Tip20-8p revealed that some regions, particularly the N-terminal domain and parts of the stalk region, were more flexible in the mutant protein, consistent with its increased susceptibility to proteolysis. Both Tip20-5p and Tip20-8p mutants prevented proper ER trans-SNARE complex assembly in vitro. Moreover, Tip20p mutant proteins disturbed the interaction between Dsl1p and the coatomer coat complex, indicating that the Dsl1p-coatomer interaction could be stabilized or regulated by Tip20p. We provide evidence for a direct role of the Dsl1 complex, in particular Tip20p, in the formation and stabilization of ER SNARE complexes.