Create TOR1 and TOR2 Have Distinct Locations in Live Cells
TOR is a structurally and functionally conserved Ser/Thr kinase found in two multiprotein complexes that regulate many cellular processes to control cell growth. Although extensively studied, the localization of TOR is still ambiguous, possibly because endogenous TOR in live cells has not been examined. Here, we examined the localization of green fluorescent protein (GFP) tagged, endogenous TOR1 and TOR2 in live S. cerevisiae cells. A DNA cassette encoding three copies of green fluorescent protein (3XGFP) was inserted in the TOR1 gene (at codon D330) or the TOR2 gene (at codon N321). The TORs were tagged internally because TOR1 or TOR2 tagged at the N or C terminus was not functional. The TOR1D330-3XGFP strain was not hypersensitive to rapamycin, was not cold sensitive, and was not resistant to manganese toxicity caused by the loss of Pmr1, all indications that TOR1-3XGFP was expressed and functional. TOR2-3XGFP was functional, as TOR2 is an essential gene and TOR2N321-3XGFP haploid cells were viable. Thus, TOR1 and TOR2 retain function after the insertion of 748 amino acids in a variable region of their noncatalytic domain. The localization patterns of TOR1-3XGFP and TOR2-3XGFP were documented by imaging of live cells. TOR1-3XGFP was diffusely cytoplasmic and concentrated near the vacuolar membrane. The TOR2-3XGFP signal was cytoplasmic but predominately in dots at the plasma membrane. Thus, TOR1 and TOR2 have distinct localization patterns, consistent with the regulation of cellular processes as part of two different complexes.