Create Overview of workflows and data approaches in bioproduction – an industry perspective
Downstream processing is an integral part of the production process of biopharmaceuticals and contributes quite significantly to the overall productivity and product quality, as well as to processing costs. Melanie Diefenbacher of Genedata provides a comprehensive overview of downstream processes, highlighting several challenges and the importance of investing into automation and digitalisation to ensure efficient processing.