Create Ypt1p is essential for retrograde Golgi-ER transport and for Golgi maintenance in S. cerevisiae
The small GTPase Ypt1p of the Rab family is required for docking of ER-derived transport vesicles with the Golgi prior to fusion. However, the identity of the Rab protein that mediates docking of Golgi-derived COPI vesicles with the ER in retrograde transport remains elusive. Here, we show that in yeast Ypt1p is essential for retrograde transport from the Golgi to the ER. Retrieval of gpαF-HDEL (glycolylated pro-α-factor with an HDEL tag at the C-terminus) was blocked in Δypt1/SLY1-20 membranes at the restrictive temperature in vitro. Moreover, Ypt1p and the ER-resident t-SNARE Ufe1p interact genetically and biochemically, indicating a role for Ypt1p in consumption of COPI vesicles at the ER. Ypt1p is also essential for the maintenance of the morphology and the protein composition of the Golgi. Interestingly, the concentrations of the Golgi enzymes Anp1p and Mnn1p, the cargo protein Emp47p and the v-SNARE Sec22p were all substantially reduced in Golgi from a Δypt1/SLY1-20 strain as compared with wild-type Golgi, while the concentration of Arf1p and of coatomer were mildly affected. Finally, COPI vesicles generated from Δypt1/SLY1-20 Golgi membranes in vitro were depleted of Emp47p and Sec22p. These data demonstrate that Ypt1p plays an essential role in retrograde transport from the Golgi to the ER.